Break up test_tune_restore
